ADD the first iteration of the RPS
I am writing the code 'blind' as in did not read the instruction until the end.
The code was very rough, but met all the condition mentioned in the practice reqs.
THE FILE is a BLANK HTML and BLANK CSS as focus are all in writing functions and such in JS.
THUS OUTPUT only on the javascript only.

I also opt the draw value, as a bit too lazy doing more while coding if result is draw.

much improvement can be done, but overall pretty happy on the project.

less likely to edit the page, perchance opt to press on the lesson.